+++ R/sppt_diff.R
@@ -1,32 +1,24 @@
-#' @import sp
+#' @importFrom sp over
+#' @importFrom dplyr group_by summarise
 NULL
 
-#' @import exact2x2
-NULL
-
-#' Performs a Spatial Point Pattern Test (SPPT) following Wheeler et al. (2018)
+#' Performs a Spatial Point Pattern Test following Wheeler et al. (2018)
 #'
 #' As opposed to the traditional SPPT, this test calculates the difference in
-#' the proportions using either a Chi-square proportions test or using Fisher's
-#' exact test, the p-values are then adjusted for multiple comparisons. See
-#' https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3111822 for a description
-#' and example. \cr
-#' If test = 'Fisher', Fisher's exact test is used but with adjusted confidence
-#' intervals (method = 'minlike'). See the R package exact2x2 for details
-#' (https://cran.r-project.org/package=exact2x2)
+#' the proportions using either a variety of Chi-square proportions tests or using
+#' Fisher's exact test, and then the p-values are adjusted for multiple comparisons.
+#' See https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3111822 for a description
+#' and example.
 #'
 #' @param p1.sp           the first points spatialobject, order does not matter
 #' @param p2.sp           the second points spatialobject, order does not matter
 #' @param uoa.sp          the units of analysis spatial object
 #' @param conf_level      confidence interval, default = 95
 #' @param test            test to conduct, currently either 'Yates' or 'Fisher', default Yates
-#' @param tsmethod        if test equals 'Fisher', then exact2x2:exact2x2() is used, with default 'minlike'.
-#'                        With Yates's test, method is ignored.
-#' @param adj             method to adjust p-values (as per p.adjust), default "BY"
-#' @return Returns a spatial polygons data frame object that has fields related to the test
-#'         calculated. When test equals \code{Fisher}, the lower and upper confidence intervals are
-#'         for the odds ratio. When test equals \code{Yates}, the lower and upper confidence intervals
-#'         are for the difference in proportions
+#' @param tsmethod        test to conduct, choice of "Chi2_Nmin1", "Chi2",
+#'                        "Yates", or "Fisher", default = "Chi2_Nmin1"
+#' @param adj             method to adjust p-values (as per p.adjust), default "BY". To turn off, set to "none".
+#' @return Returns uoa.sp (a spatial polygons data frame) including sppt_diff outcomes.
 #' @examples
 #' # Plot areas, base points data, and test points data
 #' plot(areas.sp)
@@ -34,7 +26,7 @@ NULL
 #' text(coordinates(points1.sp), label = points1.sp$ID, col="blue")
 #' text(coordinates(points2.sp), label = points2.sp$ID, col="red")
 #'
-#' output <- sppt_diff(points1.sp, points2.sp, areas.sp, test="Yates")
+#' output <- sppt_diff(points1.sp, points2.sp, areas.sp, test="Chi2_Nmin1")
 #' summary.sppt(output)
 #' output@data
 #'
@@ -52,15 +44,18 @@ NULL
 #' # The sppt_diff() outcomes without correction for multiple comparisons
 #' # are close to sppt_boot() outcomes:
 #' set.seed(9866)
-#' myoutput1 <- sppt_diff(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp, test = "Fisher", adj = "none")
+#' myoutput1 <- sppt_diff(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp, test = "Chi2_Nmin1", adj = "none")
 #' summary.sppt(myoutput1)
 #'
 #' set.seed(9866)
-#' myoutput2 <- sppt_boot(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp, nsamples = 200)
+#' myoutput2 <- sppt_boot(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp)
 #' summary.sppt(myoutput2)
 #'
 #' @export
-sppt_diff <- function(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Yates", tsmethod = "minlike", adj = "BY"){
+sppt_diff <- function(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Chi2_Nmin1", adj = "BY"){
+
+  # Check Test
+  if(!test %in% c("Chi2_Nmin1", "Chi2", "Yates", "Fisher")) stop( paste("Your test argument of", test,"is not valid.") )
 
   #################################
   # Read-in Units of Analysis
@@ -80,11 +75,13 @@ sppt_diff <- function(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Yates", tsm
   basedata_over_results <- data.frame(point_id = 1:length(p1), uoa_id = sp::over(p1, uoa)$uoa_id)
 
   # number of points per unit (only for units that have any points)
-  npoints_per_uoa <- stats::aggregate(. ~ uoa_id, data = basedata_over_results, FUN=length)
+  npoints_per_uoa <- dplyr::group_by(basedata_over_results, uoa_id)
+  npoints_per_uoa <- dplyr::summarize(npoints_per_uoa, npoints = n())
+  npoints_per_uoa <- as.data.frame(npoints_per_uoa)
 
   # fill outcome data.frame
   outcome <- data.frame(uoa_id = uoa$uoa_id)
-  outcome$nevents.b <- npoints_per_uoa[match(outcome$uoa_id, npoints_per_uoa$uoa_id), "point_id"]
+  outcome$nevents.b <- npoints_per_uoa[match(outcome$uoa_id, npoints_per_uoa$uoa_id), "npoints"]
   outcome$nevents.b[is.na(outcome$nevents.b)] <- 0
   # add percentages:
   outcome$perc.b <- with(outcome, nevents.b/sum(nevents.b)*100)
@@ -102,10 +99,12 @@ sppt_diff <- function(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Yates", tsm
   testdata_over_results <- data.frame(point_id = 1:length(p2), uoa_id = sp::over(p2, uoa)$uoa_id)
 
   # number of points per unit (only for units that have any points)
-  npoints_per_uoa <- stats::aggregate(. ~ uoa_id, data = testdata_over_results, FUN=length)
+  npoints_per_uoa <- dplyr::group_by(testdata_over_results, uoa_id)
+  npoints_per_uoa <- dplyr::summarize(npoints_per_uoa, npoints = n())
+  npoints_per_uoa <- as.data.frame(npoints_per_uoa)
 
   # base data outcome data.frame
-  outcome$nevents.t <- npoints_per_uoa[match(outcome$uoa_id, npoints_per_uoa$uoa_id), "point_id"]
+  outcome$nevents.t <- npoints_per_uoa[match(outcome$uoa_id, npoints_per_uoa$uoa_id), "npoints"]
   outcome$nevents.t[is.na(outcome$nevents.t)] <- 0
   # add percentages:
   outcome$perc.t <- with(outcome, nevents.t/sum(nevents.t)*100)
@@ -113,58 +112,34 @@ sppt_diff <- function(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Yates", tsm
   outcome$tot.t <- with(outcome, sum(nevents.t))
 
   #################################
-  # calculating p-value and confidence interval of the differences
+  # calculating p-value of the differences
   #################################
 
   outcome$diff_perc <- with(outcome, perc.b - perc.t)
   outcome$p.value <- -1
-  outcome$lower <- NA
-  outcome$upper <- NA
-  outcome$SIndex <- NA
 
-  # calculating p-values and confidence intervals
-  # can likely be improved via vectorization, also should add in Agresti-Coffo method
-  if (test=="Yates"){
-    for (i in outcome$uoa_id){
-	    ro <- outcome[outcome$uoa_id == i,]
-	    da <- matrix(c(ro$nevents.b, ro$nevents.t, ro$tot.b - ro$nevents.b, ro$tot.t - ro$nevents.t), ncol=2)
-	    res <- suppressWarnings(prop.test(da, conf.level = conf_level/100, correct = TRUE))
-	    outcome[outcome$uoa_id == i, c("p.value")] <- res$p.value
-      outcome[outcome$uoa_id == i, c("lower")] <- res$conf.int[1]*100
-      outcome[outcome$uoa_id == i, c("upper")] <- res$conf.int[2]*100
+  outcome$p.value <- sapply(outcome$uoa_id, function(x){
+    ro <- outcome[outcome$uoa_id == x, ]
+    da <- matrix(c(ro$nevents.b, ro$nevents.t, ro$tot.b - ro$nevents.b, ro$tot.t - ro$nevents.t), ncol = 2)
 
-      # Calculate S Index within loop
-      if (res$conf.int[1] > 0){
-        outcome[outcome$uoa_id == i,c("SIndex")] <- 1
-      } else if (res$conf.int[2] < 0){
-        outcome[outcome$uoa_id == i,c("SIndex")] <- -1
-      } else {
-        outcome[outcome$uoa_id == i,c("SIndex")] <- 0
-      }
-	  }
-  } else if (test=="Fisher"){
-    for (i in outcome$uoa_id){
-  	  ro <- outcome[outcome$uoa_id == i,]
-  	  da <- matrix(c(ro$nevents.b, ro$nevents.t, ro$tot.b - ro$nevents.b, ro$tot.t - ro$nevents.t), ncol=2)
-  	  res <- exact2x2::exact2x2(da, conf.level = conf_level/100, tsmethod = tsmethod)
-  	  outcome[outcome$uoa_id == i, c("p.value")] <- res$p.value
-
-  	  #These confidence intervals are for the odds ratio - not for the difference in percentages
-      outcome[outcome$uoa_id == i,c("lower")] <- res$conf.int[1]
-      outcome[outcome$uoa_id == i,c("upper")] <- res$conf.int[2]
-
-      #Since it is odds ratio, check above/below 1, not 0
-      if (res$conf.int[1] > 1){
-        outcome[outcome$uoa_id == i,c("SIndex")] <- 1
-      } else if (res$conf.int[2] < 1){
-        outcome[outcome$uoa_id == i,c("SIndex")] <- -1
-      } else {
-        outcome[outcome$uoa_id == i,c("SIndex")] <- 0
-      }
+    if (test == "Chi2_Nmin1"){
+      r <- suppressWarnings(as.numeric(chisq.test(da, correct = FALSE)$statistic))
+      N <- sum(da)
+      pval <- pchisq(r*(N-1)/N, 1, lower.tail = FALSE)
+    } else if (test %in% c("Chi2", "Yates")){
+      res <- suppressWarnings(prop.test(da, conf.level = conf_level/100, correct = ifelse(test == "Yates", TRUE, FALSE)))
+      pval <- res$p.value
+    } else if (test == "Fisher") {
+      res <- suppressWarnings(fisher.test(da, conf.level = conf_level/100))
+      pval <- res$p.value
     }
-  } else {
-    stop( paste("Your test argument of", test,"is not valid, please specify either 'Yates' or 'Fisher'") )
-  }
+
+    return(pval)
+  })
+
+  outcome$SIndex <- 0
+  outcome$SIndex[(outcome$p.value < 1 - conf_level/100) & (outcome$perc.b > outcome$perc.t)] <- 1
+  outcome$SIndex[(outcome$p.value < 1 - conf_level/100) & (outcome$perc.b < outcome$perc.t)] <- -1
 
   # doing the multiple comparison adjustment to the p-values
   outcome$p.adjusted <- p.adjust(p = outcome$p.value, method = adj)